move test deck utils into libs/utils
